Add roomStatus store
